LayoutTests/fast/dom/Geolocation/notimer-after-unload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: line 22: Use of window.alert is not allowed while unloading a page.
You should have seen one unload alert appear.
